test(llm): the memory gauge gets a process of its own
